nfc: create a bootstrap service
Since registration of NfcService as a manager service happens in the first run of the Nfc app,
and since the SystemUi makes use of this manager service through the NfcTile, NfcManager
can lead into a null reference for the NfcAdapter because the NfcService isn't registered
when the SystemUi starts (this condition depends on the device speed to bring up the call
to the apps).
This change adds a noop service that allow to be called by SystemServer before start the
systemui and ensure that NfcService is available before apps start using NfcManager.

Additional info: http://review.cyanogenmod.org/#/c/108411

TICKET: CYNGNOS-848
Change-Id: I052ab0cfc8b62fbc3798ab3c0d0eccfad01419ff
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>